+++ piecrust-uplink/src/types.rs
@@ -35,6 +35,31 @@ pub struct Event {
     pub data: Vec<u8>,
 }
 
+/// An event emitted by a contract, enriched with execution metadata.
+#[derive(
+    Debug,
+    Clone,
+    PartialEq,
+    Eq,
+    PartialOrd,
+    Ord,
+    Hash,
+    Archive,
+    Serialize,
+    Deserialize,
+)]
+#[archive_attr(derive(CheckBytes))]
+pub struct EnrichedEvent {
+    pub canonical_event: Event,
+    pub reverted: bool,
+}
+
+impl From<EnrichedEvent> for Event {
+    fn from(event: EnrichedEvent) -> Self {
+        event.canonical_event
+    }
+}
+
 /// Type with `rkyv` serialization capabilities for specific types.
 pub type StandardBufSerializer<'a> = CompositeSerializer<
     BufferSerializer<&'a mut [u8]>,